Move printers to individual modules
