+++ baleen-printer/README.md
@@ -1,0 +1,3 @@
+# Baleen Printers
+
+Light-weight [Printer](../baleen/src/main/kotlin/com/shoprunner/baleen/printer/Printer.kt) implementations.
