+++ baleen-printer/baleen-printer-csv/build.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    id("baleen.project-conventions")
+    id("baleen.publish")
+}
+
+dependencies {
+    implementation(project(":baleen"))
+}
