+++ baleen-printer/baleen-printer-csv/src/test/kotlin/com/shoprunner/baleen/printer/CsvPrinterTest.kt
@@ -1,4 +1,4 @@
-package com.shoprunner.baleen.printer.jvm
+package com.shoprunner.baleen.printer
 
 import com.shoprunner.baleen.ValidationError
 import com.shoprunner.baleen.ValidationInfo
